requirements.txt
accelerate
aiohappyeyeballs
aiohttp
aiosignal
aiosqlite
alembic
altair
annotated-types
anyio
apprise
asgi-lifespan
asttokens
asyncpg
attrs
blinker
cachetools
certifi
cffi
charset-normalizer
click
cloudpickle
colorama
comm
contourpy
coolname
cryptography
cycler
datasets
dateparser
debugpy
decorator
detoxify
dill
docker
exceptiongroup
executing
fastapi
filelock
fonttools
frozenlist
fsspec
git-filter-repo
gitdb
GitPython
graphviz
greenlet
griffe
grpclib
h11
h2
hf-xet
hpack
httpcore
httpx
huggingface-hub
humanize
hyperframe
idna
importlib_metadata
ipykernel
ipython
ipython_pygments_lexers
jedi
Jinja2
jinja2-humanize-extension
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
kiwisolver
llvmlite
Mako
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
modal
mpmath
multidict
multiprocess
narwhals
nest-asyncio
networkx
numba
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-cufile-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cusparselt-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
oauthlib
opentelemetry-api
orjson
packaging
pandas
parso
pathspec
peft
pendulum
pexpect
pillow
platformdirs
praw
prawcore
prefect
prometheus_client
prompt_toolkit
propcache
protobuf
psutil
psycopg2-binary
ptyprocess
pure_eval
pyarrow
pycparser
pydantic
pydantic-extra-types
pydantic-settings
pydantic_core
pydeck
Pygments
pynndescent
pyparsing
python-dateutil
python-dotenv
python-slugify
python-socks
pytz
PyYAML
pyzmq
readchar
referencing
regex
requests
requests-oauthlib
rfc3339-validator
rich
rpds-py
ruamel.yaml
ruamel.yaml.clib
safetensors
scikit-learn
scipy
sentence-transformers
sentencepiece
setuptools
shellingham
sigtools
six
smmap
sniffio
SQLAlchemy
stack-data
starlette
streamlit
sympy
synchronicity
tenacity
text-unidecode
threadpoolctl
tokenizers
toml
torch
tornado
tqdm
traitlets
transformers
triton
typer
types-certifi
types-toml
typing-inspection
typing_extensions
tzdata
tzlocal
ujson
umap-learn
update-checker
urllib3
uv
uvicorn
watchdog
watchfiles
wcwidth
websocket-client
websockets
xxhash
yarl
zipp
pytest